move audit log stuff to its own pkg
