+++ pkg/api/projects.go
@@ -29,6 +29,7 @@ import (
 
 	"github.com/gorilla/mux"
 	"github.com/sapcc/go-bits/respondwith"
+	"github.com/sapcc/limes/pkg/audit"
 	"github.com/sapcc/limes/pkg/collector"
 	"github.com/sapcc/limes/pkg/db"
 	"github.com/sapcc/limes/pkg/limes"
@@ -244,7 +245,7 @@ func (p *v1Provider) PutProject(w http.ResponseWriter, r *http.Request) {
 	servicesToUpdate := make(map[string]bool)
 	var errors []string
 
-	var auditTrail util.AuditTrail
+	var auditTrail audit.AuditTrail
 	for _, srv := range services {
 		resourceQuotas, exists := serviceQuotas[srv.Type]
 		if !exists {
@@ -285,7 +286,7 @@ func (p *v1Provider) PutProject(w http.ResponseWriter, r *http.Request) {
 			//would contain only identical pointers)
 			res := res
 
-			auditEvent := util.NewAuditEvent(token, r, requestTime, dbProject.UUID, srv.Type, res.Name, res.Quota, newQuota)
+			auditEvent := audit.NewAuditEvent(token, r, requestTime, dbProject.UUID, srv.Type, res.Name, res.Quota, newQuota)
 			auditTrail.Add(auditEvent)
 
 			res.Quota = newQuota
