+++ pkg/audit/log.go
@@ -17,7 +17,7 @@
 *
 *******************************************************************************/
 
-package util
+package audit
 
 import (
 	"encoding/json"
