+++ pkg/audit/logmiddleware.go
@@ -17,7 +17,7 @@
 *
 *******************************************************************************/
 
-package util
+package audit
 
 import (
 	"bytes"
